+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * *BREAKING* Use `convertToSlipData` and supporting functions from `stripes-util`. Refs UIREQ-1263.
 * Replace moment with day.js. Refs UIREQ-1291.
+* [Retrieval SP filter] Move related code to make the filter enabled for ecs/non-ecs UI. Refs UIREQ-1294.
 
 ## [12.0.1] (https://github.com/folio-org/ui-requests/tree/v12.0.1) (2025-03-21)
 [Full Changelog](https://github.com/folio-org/ui-requests/compare/v12.0.0...v12.0.1)